backend-py/requirements.txt
fastapi[all] 
uvicorn 
pydantic 
python-dotenv 
passlib[bcrypt] 
python-jose[cryptography] 
PyJWT 
numpy 
pandas 
requests
bcrypt
pymongo
typing
datetime
uuid